Set RACK_ENV to env
This is unfortunate but we need to live with it for now. Currently
container in dry-web sets its config.env based on this env var.
Eventually we should move env setting into its own container variable,
as it should not be part of the config because we can't enforce it.