Introduce Snowflakes::Application
This subclasses Dry::System::Container and is intended to act as a single instance representing the overall "application", which has access to the multiple "slices" that comprise it (previously the slightly clumsy "sub-apps" in our parlance).

This single application is made available via `Snowflakes.application`, which helps with simplifying the CLI creation, and will (hopefully!) help with managing the test suite helpers.